+++ Distroir.CustomSDKLauncher.Core/Managers/Async/AsyncManager.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using Distroir.CustomSDKLauncher.Core.Managers.Async.AsyncContentSerializers;
+using Distroir.CustomSDKLauncher.Core.Managers.Async.Serializers;
 using Distroir.CustomSDKLauncher.Core.Utilities;
 
 namespace Distroir.CustomSDKLauncher.Core.Managers.Async
