+++ Distroir.CustomSDKLauncher.Core/Managers/Async/Serializers/ISerializer.cs
@@ -1,4 +1,4 @@
-namespace Distroir.CustomSDKLauncher.Core.Managers.Async.AsyncContentSerializers
+namespace Distroir.CustomSDKLauncher.Core.Managers.Async.Serializers
 {
     public interface ISerializer<T>
     {
